Added basic logic for GlyphEntry

+++ src/Configuration/SFrame.cs
@@ -6,8 +6,8 @@ namespace Glyphy.Configuration
     public struct SFrame
     {
         public uint Frame { get; set; }
-        public uint TransitionTime { get; set; }
-        public uint Duration { get; set; }
+        public double TransitionTime { get; set; }
+        public double Duration { get; set; }
         public Dictionary<EAddressable, SLEDValue> Values { get; set; }
     }
 }
